db extract watch

+++ watch.py
@@ -1,8 +1,15 @@
+import argparse
 from shlex import quote
 from rich import inspect, print
-from db import con
+from db import sqlite_con
 from utils import cmd
 import os
+
+parser = argparse.ArgumentParser()
+parser.add_argument("db", default="./videos.db")
+parser.add_argument("-keep", "--keep", action="store_true")
+args = parser.parse_args()
+con = sqlite_con(args.db)
 
 next_video = dict(
     con.execute(
@@ -24,9 +31,10 @@ limit 1
 )["filename"]
 
 print(next_video)
-if os.path.exists(next_video):
-    cmd(f"mpv --quiet {quote(next_video)} --fs")
-    cmd(f"trash-put {quote(next_video)}")
+if not args.keep:
+    if os.path.exists(next_video):
+        cmd(f"mpv --quiet {quote(next_video)} --fs")
+        cmd(f"trash-put {quote(next_video)}")
 
-con.execute("delete from videos where filename = ?", (next_video,))
-con.commit()
+    con.execute("delete from videos where filename = ?", (next_video,))
+    con.commit()
